+++ OUDS/Core/Themes/Orange/Sources/Providers/ComponentTokens/OrangeThemeBreadcrumbComponentTokensProvider.swift
@@ -23,8 +23,7 @@ import OUDSTokensSemantic
 /// all tokens to the users. It helps users to override some of the tokens and assign them to an `OUDSTheme` implementation to use.
 /// Custom themes can use subclass of ``OrangeThemeBreadcrumbComponentTokensProvider`` and apply the provider they need.
 /// It implements also the protocol `BreadcrumbComponentTokens` so as to expose the component tokens for *breadcrumb* through any `OUDSTheme`.
-/// *Breadcrumb* components tokens are defined with raw and semantic tokens of sizes (from `AllSizeSemanticTokensProvider`),
-/// colors (from `AllColorSemanticTokensProvider`) and spaces (from `AllSpaceSemanticTokensProvider`).
+/// *Breadcrumb* components tokens are defined with raw and semantic tokens of sizes (from `AllSizeSemanticTokensProvider`)  and spaces (from `AllSpaceSemanticTokensProvider`).
 ///
 /// ```swift
 ///     // Define your own provider for breadcrumb component tokens
@@ -33,11 +32,9 @@ import OUDSTokensSemantic
 ///
 ///         // Then override the breadcrumb component tokens you want.
 ///
-///         override var breadcrumbSizeMinHeightMedium: SizeSemanticToken { DimensionRawTokens.dimension550 }
+///         override var breadcrumbSizeArrow: SizeSemanticToken { DimensionRawTokens.dimension550 }
 ///
-///         override var breadcrumbColorContentEnabled: MultipleColorSemanticTokens { colors.colorDecorativeSkinTint200 }
-///
-///         override var breadcrumbSpaceColumnGapLevelsMedium: SpaceSemanticToken { spaces.spaceColumnGapShorter }
+///         override var breadcrumbSpaceColumnGapArrow: SpaceSemanticToken { spaces.spaceColumnGapShorter }
 ///
 ///         // ...
 ///     }
@@ -73,23 +70,17 @@ open class OrangeThemeBreadcrumbComponentTokensProvider: AllBreadcrumbComponentT
     /// Provider of size semantic tokens to use for link sizes
     public let sizes: AllSizeSemanticTokensProvider
 
-    /// Provider of color semantic tokens to use for link colors
-    public let colors: AllColorSemanticTokensProvider
-
     /// Provider of spaces semantic tokens to use for link spaces
     public let spaces: AllSpaceSemanticTokensProvider
 
     /// Defines a provider of component tokens dedicated to `OUDSLink`
     /// - Parameters:
     ///    - sizes: Provider for size semantic tokens
-    ///    - colors: Provider for color semantic tokens
     ///    - spaces: Provider for space semantic tokens
     public init(sizes: AllSizeSemanticTokensProvider,
-                colors: AllColorSemanticTokensProvider,
                 spaces: AllSpaceSemanticTokensProvider) {
         OL.debug("Init of OrangeThemeBreadcrumbComponentTokensProvider")
         self.sizes = sizes
-        self.colors = colors
         self.spaces = spaces
     }