+++ OUDS/Core/Themes/Orange/Sources/Values/SemanticTokens/OrangeTheme+BorderSemanticTokens.swift
@@ -16,6 +16,7 @@ import OUDSTokensRaw
 import OUDSTokensSemantic
 
 extension OrangeThemeBorderSemanticTokensProvider: BorderSemanticTokens {
+<<<<<<< HEAD
     @objc open var borderRadiusDefault: BorderRadiusSemanticToken { BorderRawTokens.borderRadius0 }
     @objc open var borderRadiusMedium: BorderRadiusSemanticToken { BorderRawTokens.borderRadius150 }
     @objc open var borderRadiusNone: BorderRadiusSemanticToken { BorderRawTokens.borderRadius0 }
@@ -27,8 +28,22 @@ extension OrangeThemeBorderSemanticTokensProvider: BorderSemanticTokens {
     @objc open var borderWidthFocus: BorderWidthSemanticToken { BorderRawTokens.borderWidth125 }
     @objc open var borderWidthFocusInset: BorderWidthSemanticToken { BorderRawTokens.borderWidth50 }
     @objc open var borderWidthMedium: BorderWidthSemanticToken { BorderRawTokens.borderWidth50 }
+=======
+>>>>>>> 4505fe428 (refactor(🤖): partial integration of tokens version 0.6.0 (#458))
     @objc open var borderWidthNone: BorderWidthSemanticToken { BorderRawTokens.borderWidth0 }
+    @objc open var borderWidthDefault: BorderWidthSemanticToken { BorderRawTokens.borderWidth25 }
+    @objc open var borderWidthThin: BorderWidthSemanticToken { BorderRawTokens.borderWidth25 }
+    @objc open var borderWidthMedium: BorderWidthSemanticToken { BorderRawTokens.borderWidth50 }
     @objc open var borderWidthThick: BorderWidthSemanticToken { BorderRawTokens.borderWidth75 }
     @objc open var borderWidthThicker: BorderWidthSemanticToken { BorderRawTokens.borderWidth100 }
-    @objc open var borderWidthThin: BorderWidthSemanticToken { BorderRawTokens.borderWidth25 }
+    @objc open var borderWidthFocusInset: BorderWidthSemanticToken { BorderRawTokens.borderWidth50 }
+    @objc open var borderWidthFocus: BorderWidthSemanticToken { BorderRawTokens.borderWidth125 }
+    @objc open var borderRadiusNone: BorderRadiusSemanticToken { BorderRawTokens.borderRadius0 }
+    @objc open var borderRadiusDefault: BorderRadiusSemanticToken { BorderRawTokens.borderRadius0 }
+    @objc open var borderRadiusShort: BorderRadiusSemanticToken { BorderRawTokens.borderRadius75 }
+    @objc open var borderRadiusMedium: BorderRadiusSemanticToken { BorderRawTokens.borderRadius150 }
+    @objc open var borderRadiusTall: BorderRadiusSemanticToken { BorderRawTokens.borderRadius300 }
+    @objc open var borderRadiusPill: BorderRadiusSemanticToken { BorderRawTokens.borderRadius9999 }
+    @objc open var borderStyleDefault: BorderStyleSemanticToken { BorderRawTokens.borderStyleSolid }
+    @objc open var borderStyleDrag: BorderStyleSemanticToken { BorderRawTokens.borderStyleDashed }
 }