+++ OUDS/Core/Themes/Orange/Tests/Values/ComponentTokens/TestThemeOverrideOfChipComponentTokens.swift
@@ -38,15 +38,24 @@ final class TestThemeOverrideOfChipComponentTokens: XCTestCase {
         XCTAssertTrue(inheritedTheme.chip.chipSizeMinHeight == MockThemeChipComponentTokenProvider.mockThemeChipSize)
     }
 
+    func testInheritedThemeCanOverrideChipComponentTokenSizeMinWidth() throws {
+        XCTAssertNotEqual(inheritedTheme.chip.chipSizeMinWidth, abstractTheme.chip.chipSizeMinWidth)
+        XCTAssertTrue(inheritedTheme.chip.chipSizeMinWidth == MockThemeChipComponentTokenProvider.mockThemeChipSize)
+    }
+
     func testInheritedThemeCanOverrideChipComponentTokenSizeIcon() throws {
         XCTAssertNotEqual(inheritedTheme.chip.chipSizeIcon, abstractTheme.chip.chipSizeIcon)
         XCTAssertTrue(inheritedTheme.chip.chipSizeIcon == MockThemeChipComponentTokenProvider.mockThemeChipSize)
     }
 
+<<<<<<< HEAD
     func testInheritedThemeCanOverrideChipComponentTokenBorderRadius() throws {
         XCTAssertNotEqual(inheritedTheme.chip.chipBorderRadius, abstractTheme.chip.chipBorderRadius)
         XCTAssertTrue(inheritedTheme.chip.chipBorderRadius == MockThemeChipComponentTokenProvider.mockThemeChipBorderRadius)
     }
+=======
+    // MARK: - Borders
+>>>>>>> 4505fe428 (refactor(🤖): partial integration of tokens version 0.6.0 (#458))
 
     func testInheritedThemeCanOverrideChipComponentTokenBorderWidthDefault() throws {
         XCTAssertNotEqual(inheritedTheme.chip.chipBorderWidthDefault, abstractTheme.chip.chipBorderWidthDefault)
@@ -58,21 +67,64 @@ final class TestThemeOverrideOfChipComponentTokens: XCTestCase {
         XCTAssertTrue(inheritedTheme.chip.chipBorderWidthDefaultInteraction == MockThemeChipComponentTokenProvider.mockThemeChipBorderWidth)
     }
 
+<<<<<<< HEAD
     func testInheritedThemeCanOverrideChipComponentTokenColorBgDisabled() throws {
         XCTAssertNotEqual(inheritedTheme.chip.chipColorBgDisabled, abstractTheme.chip.chipColorBgDisabled)
         XCTAssertTrue(inheritedTheme.chip.chipColorBgDisabled == MockThemeChipComponentTokenProvider.mockThemeChipColor)
     }
 
+=======
+    func testInheritedThemeCanOverrideChipComponentTokenBorderWidthSelected() throws {
+        XCTAssertNotEqual(inheritedTheme.chip.chipBorderWidthSelected, abstractTheme.chip.chipBorderWidthSelected)
+        XCTAssertTrue(inheritedTheme.chip.chipBorderWidthSelected == MockThemeChipComponentTokenProvider.mockThemeChipBorderWidth)
+    }
+
+    func testInheritedThemeCanOverrideChipComponentTokenBorderRadiusPill() throws {
+        XCTAssertNotEqual(inheritedTheme.chip.chipBorderRadiusPill, abstractTheme.chip.chipBorderRadiusPill)
+        XCTAssertTrue(inheritedTheme.chip.chipBorderRadiusPill == MockThemeChipComponentTokenProvider.mockThemeChipBorderRadius)
+    }
+
+    // MARK: - Colors
+
+    func testInheritedThemeCanOverrideChipComponentTokenColorContentEnabled() throws {
+        XCTAssertNotEqual(inheritedTheme.chip.chipColorContentEnabled, abstractTheme.chip.chipColorContentEnabled)
+        XCTAssertTrue(inheritedTheme.chip.chipColorContentEnabled == MockThemeChipComponentTokenProvider.mockThemeChipColor)
+    }
+
+    func testInheritedThemeCanOverrideChipComponentTokenColorContentHover() throws {
+        XCTAssertNotEqual(inheritedTheme.chip.chipColorContentHover, abstractTheme.chip.chipColorContentHover)
+        XCTAssertTrue(inheritedTheme.chip.chipColorContentHover == MockThemeChipComponentTokenProvider.mockThemeChipColor)
+    }
+
+    func testInheritedThemeCanOverrideChipComponentTokenColorContentPressed() throws {
+        XCTAssertNotEqual(inheritedTheme.chip.chipColorContentPressed, abstractTheme.chip.chipColorContentPressed)
+        XCTAssertTrue(inheritedTheme.chip.chipColorContentPressed == MockThemeChipComponentTokenProvider.mockThemeChipColor)
+    }
+
+    func testInheritedThemeCanOverrideChipComponentTokenColorContentDisabled() throws {
+        XCTAssertNotEqual(inheritedTheme.chip.chipColorContentDisabled, abstractTheme.chip.chipColorContentDisabled)
+        XCTAssertTrue(inheritedTheme.chip.chipColorContentDisabled == MockThemeChipComponentTokenProvider.mockThemeChipColor)
+    }
+
+    func testInheritedThemeCanOverrideChipComponentTokenColorContentFocus() throws {
+        XCTAssertNotEqual(inheritedTheme.chip.chipColorContentFocus, abstractTheme.chip.chipColorContentFocus)
+        XCTAssertTrue(inheritedTheme.chip.chipColorContentFocus == MockThemeChipComponentTokenProvider.mockThemeChipColor)
+    }
+
+>>>>>>> 4505fe428 (refactor(🤖): partial integration of tokens version 0.6.0 (#458))
     func testInheritedThemeCanOverrideChipComponentTokenColorBgEnabled() throws {
         XCTAssertNotEqual(inheritedTheme.chip.chipColorBgEnabled, abstractTheme.chip.chipColorBgEnabled)
         XCTAssertTrue(inheritedTheme.chip.chipColorBgEnabled == MockThemeChipComponentTokenProvider.mockThemeChipColor)
     }
 
+<<<<<<< HEAD
     func testInheritedThemeCanOverrideChipComponentTokenColorBgFocus() throws {
         XCTAssertNotEqual(inheritedTheme.chip.chipColorBgFocus, abstractTheme.chip.chipColorBgFocus)
         XCTAssertTrue(inheritedTheme.chip.chipColorBgFocus == MockThemeChipComponentTokenProvider.mockThemeChipColor)
     }
 
+=======
+>>>>>>> 4505fe428 (refactor(🤖): partial integration of tokens version 0.6.0 (#458))
     func testInheritedThemeCanOverrideChipComponentTokenColorBgHover() throws {
         XCTAssertNotEqual(inheritedTheme.chip.chipColorBgHover, abstractTheme.chip.chipColorBgHover)
         XCTAssertTrue(inheritedTheme.chip.chipColorBgHover == MockThemeChipComponentTokenProvider.mockThemeChipColor)
@@ -83,9 +135,20 @@ final class TestThemeOverrideOfChipComponentTokens: XCTestCase {
         XCTAssertTrue(inheritedTheme.chip.chipColorBgPressed == MockThemeChipComponentTokenProvider.mockThemeChipColor)
     }
 
+<<<<<<< HEAD
     func testInheritedThemeCanOverrideChipComponentTokenColorBorderDisabled() throws {
         XCTAssertNotEqual(inheritedTheme.chip.chipColorBorderDisabled, abstractTheme.chip.chipColorBorderDisabled)
         XCTAssertTrue(inheritedTheme.chip.chipColorBorderDisabled == MockThemeChipComponentTokenProvider.mockThemeChipColor)
+=======
+    func testInheritedThemeCanOverrideChipComponentTokenColorBgDisabled() throws {
+        XCTAssertNotEqual(inheritedTheme.chip.chipColorBgDisabled, abstractTheme.chip.chipColorBgDisabled)
+        XCTAssertTrue(inheritedTheme.chip.chipColorBgDisabled == MockThemeChipComponentTokenProvider.mockThemeChipColor)
+    }
+
+    func testInheritedThemeCanOverrideChipComponentTokenColorBgFocus() throws {
+        XCTAssertNotEqual(inheritedTheme.chip.chipColorBgFocus, abstractTheme.chip.chipColorBgFocus)
+        XCTAssertTrue(inheritedTheme.chip.chipColorBgFocus == MockThemeChipComponentTokenProvider.mockThemeChipColor)
+>>>>>>> 4505fe428 (refactor(🤖): partial integration of tokens version 0.6.0 (#458))
     }
 
     func testInheritedThemeCanOverrideChipComponentTokenColorBorderEnabled() throws {
@@ -103,11 +166,20 @@ final class TestThemeOverrideOfChipComponentTokens: XCTestCase {
         XCTAssertTrue(inheritedTheme.chip.chipColorBorderPressed == MockThemeChipComponentTokenProvider.mockThemeChipColor)
     }
 
+<<<<<<< HEAD
+=======
+    func testInheritedThemeCanOverrideChipComponentTokenColorBorderDisabled() throws {
+        XCTAssertNotEqual(inheritedTheme.chip.chipColorBorderDisabled, abstractTheme.chip.chipColorBorderDisabled)
+        XCTAssertTrue(inheritedTheme.chip.chipColorBorderDisabled == MockThemeChipComponentTokenProvider.mockThemeChipColor)
+    }
+
+>>>>>>> 4505fe428 (refactor(🤖): partial integration of tokens version 0.6.0 (#458))
     func testInheritedThemeCanOverrideChipComponentTokenColorBorderSelected() throws {
         XCTAssertNotEqual(inheritedTheme.chip.chipColorBorderSelected, abstractTheme.chip.chipColorBorderSelected)
         XCTAssertTrue(inheritedTheme.chip.chipColorBorderSelected == MockThemeChipComponentTokenProvider.mockThemeChipColor)
     }
 
+<<<<<<< HEAD
     func testInheritedThemeCanOverrideChipComponentTokenColorContentDisabled() throws {
         XCTAssertNotEqual(inheritedTheme.chip.chipColorContentDisabled, abstractTheme.chip.chipColorContentDisabled)
         XCTAssertTrue(inheritedTheme.chip.chipColorContentDisabled == MockThemeChipComponentTokenProvider.mockThemeChipColor)
@@ -136,26 +208,23 @@ final class TestThemeOverrideOfChipComponentTokens: XCTestCase {
     func testInheritedThemeCanOverrideChipComponentTokenSpaceColumnGapColumnGap() throws {
         XCTAssertNotEqual(inheritedTheme.chip.chipSpaceColumnGapColumnGap, abstractTheme.chip.chipSpaceColumnGapColumnGap)
         XCTAssertTrue(inheritedTheme.chip.chipSpaceColumnGapColumnGap == MockThemeChipComponentTokenProvider.mockThemeChipSpace)
+=======
+    func testInheritedThemeCanOverrideChipComponentTokenColorBorderFocus() throws {
+        XCTAssertNotEqual(inheritedTheme.chip.chipColorBorderFocus, abstractTheme.chip.chipColorBorderFocus)
+        XCTAssertTrue(inheritedTheme.chip.chipColorBorderFocus == MockThemeChipComponentTokenProvider.mockThemeChipColor)
     }
 
-    func testInheritedThemeCanOverrideChipComponentTokenSpaceInsetIconOnly() throws {
-        XCTAssertNotEqual(inheritedTheme.chip.chipSpaceInsetIconOnly, abstractTheme.chip.chipSpaceInsetIconOnly)
-        XCTAssertTrue(inheritedTheme.chip.chipSpaceInsetIconOnly == MockThemeChipComponentTokenProvider.mockThemeChipSpace)
+    // MARK: - Spaces
+
+    func testInheritedThemeCanOverrideChipComponentTokenSpacePaddingInlineLabelStart() throws {
+        XCTAssertNotEqual(inheritedTheme.chip.chipSpacePaddingInlineLabelStart, abstractTheme.chip.chipSpacePaddingInlineLabelStart)
+        XCTAssertTrue(inheritedTheme.chip.chipSpacePaddingInlineLabelStart == MockThemeChipComponentTokenProvider.mockThemeChipSpace)
+>>>>>>> 4505fe428 (refactor(🤖): partial integration of tokens version 0.6.0 (#458))
     }
 
-    func testInheritedThemeCanOverrideChipComponentTokenSpacePaddingBlock() throws {
-        XCTAssertNotEqual(inheritedTheme.chip.chipSpacePaddingBlock, abstractTheme.chip.chipSpacePaddingBlock)
-        XCTAssertTrue(inheritedTheme.chip.chipSpacePaddingBlock == MockThemeChipComponentTokenProvider.mockThemeChipSpace)
-    }
-
-    func testInheritedThemeCanOverrideChipComponentTokenSpacePaddingInlineIconEnd() throws {
-        XCTAssertNotEqual(inheritedTheme.chip.chipSpacePaddingInlineIconEnd, abstractTheme.chip.chipSpacePaddingInlineIconEnd)
-        XCTAssertTrue(inheritedTheme.chip.chipSpacePaddingInlineIconEnd == MockThemeChipComponentTokenProvider.mockThemeChipSpace)
-    }
-
-    func testInheritedThemeCanOverrideChipComponentTokenSpacePaddingInlineIconNone() throws {
-        XCTAssertNotEqual(inheritedTheme.chip.chipSpacePaddingInlineIconNone, abstractTheme.chip.chipSpacePaddingInlineIconNone)
-        XCTAssertTrue(inheritedTheme.chip.chipSpacePaddingInlineIconNone == MockThemeChipComponentTokenProvider.mockThemeChipSpace)
+    func testInheritedThemeCanOverrideChipComponentTokenSpacePaddingInlineLabelEnd() throws {
+        XCTAssertNotEqual(inheritedTheme.chip.chipSpacePaddingInlineLabelEnd, abstractTheme.chip.chipSpacePaddingInlineLabelEnd)
+        XCTAssertTrue(inheritedTheme.chip.chipSpacePaddingInlineLabelEnd == MockThemeChipComponentTokenProvider.mockThemeChipSpace)
     }
 
     func testInheritedThemeCanOverrideChipComponentTokenSpacePaddingInlineIconOnly() throws {
@@ -168,14 +237,29 @@ final class TestThemeOverrideOfChipComponentTokens: XCTestCase {
         XCTAssertTrue(inheritedTheme.chip.chipSpacePaddingInlineIconStart == MockThemeChipComponentTokenProvider.mockThemeChipSpace)
     }
 
-    func testInheritedThemeCanOverrideChipComponentTokenSpacePaddingInlineLabelEnd() throws {
-        XCTAssertNotEqual(inheritedTheme.chip.chipSpacePaddingInlineLabelEnd, abstractTheme.chip.chipSpacePaddingInlineLabelEnd)
-        XCTAssertTrue(inheritedTheme.chip.chipSpacePaddingInlineLabelEnd == MockThemeChipComponentTokenProvider.mockThemeChipSpace)
+    func testInheritedThemeCanOverrideChipComponentTokenSpacePaddingInlineIconEnd() throws {
+        XCTAssertNotEqual(inheritedTheme.chip.chipSpacePaddingInlineIconEnd, abstractTheme.chip.chipSpacePaddingInlineIconEnd)
+        XCTAssertTrue(inheritedTheme.chip.chipSpacePaddingInlineIconEnd == MockThemeChipComponentTokenProvider.mockThemeChipSpace)
     }
 
-    func testInheritedThemeCanOverrideChipComponentTokenSpacePaddingInlineLabelStart() throws {
-        XCTAssertNotEqual(inheritedTheme.chip.chipSpacePaddingInlineLabelStart, abstractTheme.chip.chipSpacePaddingInlineLabelStart)
-        XCTAssertTrue(inheritedTheme.chip.chipSpacePaddingInlineLabelStart == MockThemeChipComponentTokenProvider.mockThemeChipSpace)
+    func testInheritedThemeCanOverrideChipComponentTokenSpacePaddingInlineIconNone() throws {
+        XCTAssertNotEqual(inheritedTheme.chip.chipSpacePaddingInlineIconNone, abstractTheme.chip.chipSpacePaddingInlineIconNone)
+        XCTAssertTrue(inheritedTheme.chip.chipSpacePaddingInlineIconNone == MockThemeChipComponentTokenProvider.mockThemeChipSpace)
+    }
+
+    func testInheritedThemeCanOverrideChipComponentTokenSpaceInsetIconOnly() throws {
+        XCTAssertNotEqual(inheritedTheme.chip.chipSpaceInsetIconOnly, abstractTheme.chip.chipSpaceInsetIconOnly)
+        XCTAssertTrue(inheritedTheme.chip.chipSpaceInsetIconOnly == MockThemeChipComponentTokenProvider.mockThemeChipSpace)
+    }
+
+    func testInheritedThemeCanOverrideChipComponentTokenSpacePaddingBlock() throws {
+        XCTAssertNotEqual(inheritedTheme.chip.chipSpacePaddingBlock, abstractTheme.chip.chipSpacePaddingBlock)
+        XCTAssertTrue(inheritedTheme.chip.chipSpacePaddingBlock == MockThemeChipComponentTokenProvider.mockThemeChipSpace)
+    }
+
+    func testInheritedThemeCanOverrideChipComponentTokenSpaceColumnGapColumnGap() throws {
+        XCTAssertNotEqual(inheritedTheme.chip.chipSpaceColumnGapColumnGap, abstractTheme.chip.chipSpaceColumnGapColumnGap)
+        XCTAssertTrue(inheritedTheme.chip.chipSpaceColumnGapColumnGap == MockThemeChipComponentTokenProvider.mockThemeChipSpace)
     }
 
     func testInheritedThemeCanOverrideChipComponentTokenSpacePaddingStackPaddingBlock() throws {